% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtifClusterG.R
\name{ArtifCluster}
\alias{ArtifCluster}
\title{Manually specify coordinates to display the network}
\usage{
ArtifCluster(cor = cor, nodeGroup = netClu, r = r, da = da)
}
\arguments{
\item{cor}{Correlation matrix}

\item{nodeGroup}{Classification information of network nodes}

\item{r}{Radius of each submodule}

\item{da}{The coordinates of each submodule}
}
\value{
list
}
\description{
Manually specify coordinates to display the network
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
netClu = data.frame(ID = row.names(cor),group =rep(1:3,length(row.names(cor)))[1:length(row.names(cor))] )
netClu$group = as.factor(netClu$group)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
