% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corBionetwork2.R
\name{corBionetwork}
\alias{corBionetwork}
\title{Microbial related bipartite network analysis}
\usage{
corBionetwork(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  lab = NULL,
  N = 0,
  r.threshold = 0.6,
  p.threshold = 0.05,
  label = FALSE,
  group = "Group",
  env = NULL,
  envGroup = NULL,
  method = "spearman",
  layout = "fruchtermanreingold",
  path = "./",
  fill = "Phylum",
  size = "igraph.degree",
  scale = TRUE,
  bio = TRUE,
  zipi = FALSE,
  step = 100,
  width = 20,
  height = 20,
  big = TRUE,
  select_layout = TRUE,
  layout_net = "model_maptree",
  clu_method = "cluster_fast_greedy",
  minsize = 4,
  maxsize = 14
)
}
\arguments{
\item{ps}{phyloseq Object, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N}{filter OTU tables by abundance.The defult, N=0.001, extract the top 0.001 relative abundance of OTU.}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{env}{Environmental factor index table which do network analysis with the microbiome.}

\item{envGroup}{group of env table.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{path}{save path of all of network analyse.}

\item{fill}{fill coulor of node.}

\item{size}{node size.}

\item{scale}{Whether relative abundance standardization is required.}

\item{bio}{Do you need to do a binary network.}

\item{zipi}{zipi Calculation.}

\item{step}{Random network sampling times.}

\item{width}{Save the width of the picture settings.}

\item{height}{Save the height of the picture setting.}

\item{lay}{layout which network show.}
}
\value{
list which contains OTU correlation matrix
}
\description{
Microbial related bipartite network analysis
}
\examples{
path = "./netowrk/"
data(ps)
ps16s = ps \%>\% ggClusterNet::scale_micro()
psITS = NULL
library(phyloseq)
ps.merge <- ggClusterNet::merge16S_ITS(ps16s = ps16s,
                                       psITS = NULL,
                                       N16s = 100)
map =  phyloseq::sample_data(ps.merge)
head(map)
map$Group = "one"
phyloseq::sample_data(ps.merge) <- map
data(env1)
data1 = env1
data1$id = row.names(data1)
data1 = data1 \%>\% select(id,everything())
envRDA.s = vegan::decostand(env1,"hellinger")
data1[,-1] = envRDA.s
Gru = data.frame(ID = colnames(data1)[-1],group = "env" )
head(Gru)
corBionetwork(ps = ps.merge,
                            N = 0,
                            r.threshold = 0.4, # 鐩稿叧闃堝€戼㹣
                            p.threshold = 0.05,
                            big = T,
                            group = "Group",
                            env = data1, # 鐜鎸囨爣琛ㄦ牸
                            envGroup = Gru,# 鐜鍥犲瓙鍒嗙粍鏂囦欢琛ㄦ牸
                            # layout = "fruchtermanreingold",
                            path = path,# 缁撴灉鏂囦欢瀛樺偍璺緞
                            fill = "Phylum", # 鍑哄浘鐐瑰～鍏呴鑹茬敤浠€涔堝€戼㹣
                            size = "igraph.degree", # 鍑哄浘鐐瑰ぇ灏忕敤浠€涔堟暟鎹愼㹥
                            scale = TRUE, # 鏄惁瑕佽繘琛岀浉瀵逛赴搴︽爣鍑嗗寲
                            bio = TRUE, # 鏄惁鍋氫簩鍒嗙綉缁㤼㹣
                            zipi = F, # 鏄惁璁＄畻ZIPI
                            step = 100, # 闅忔満缃戠粶鎶芥牱鐨勬鏁戼㸰
                            width = 18,
                            label = TRUE,
                            height = 10
)
}
\references{
Tao Wen#, Penghao Xie#, Shengdie Yang, Guoqing Niu, Xiaoyu Liu, Zhexu Ding, Chao Xue, Yong-Xin Liu *, Qirong Shen, Jun Yuan*
ggClusterNet: an R package for microbiome network analysis and modularity-based multiple network layouts
iMeta 2022,DOI: \url{doi: 10.1002/imt2.32}
}
\author{
Contact: Tao Wen \email{taowen@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn} yongxin liu \email{yxliu@genetics.ac.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
