% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culculate_node_axis.R
\name{culculate_node_axis}
\alias{culculate_node_axis}
\title{use layout calculated the axis a and y}
\usage{
culculate_node_axis(
  cor.matrix = cor,
  layout = "model_Gephi.2",
  seed = 1,
  group = NULL,
  model = FALSE,
  method = "cluster_fast_greedy"
)
}
\arguments{
\item{cor.matrix}{Correlation matrix}

\item{layout}{network layout algorithm in ggClusterNet}
}
\description{
filter microbiome data
}
\details{
Call the network layout algorithm for coordinate calculation
}
\examples{
data(cor)
node = culculate_node_axis(
 cor.matrix = cor,
 layout = "model_Gephi.2",
 seed = 1,
 group = NULL,
 model = FALSE,
 method = "cluster_fast_greedy")
}
