% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_culculate.R
\name{div_culculate}
\alias{div_culculate}
\title{Calculate a bipartite network for microbiome data}
\usage{
div_culculate(table = result[[3]], distance = 1.1, distance2 = 1.5,
  distance3 = 1.3, order = FALSE)
}
\arguments{
\item{table}{ven output}

\item{distance}{core distance to group}

\item{order}{TRUE or FEASE}

\item{flour}{Is the arrangement of points discrete? TRUE or FEASE would be selected}
}
\value{
list
}
\description{
Calculate a bipartite network for microbiome data
}
\examples{
data(ps)
ps_sub = filter_taxa(ps, function(x) sum(x ) > 20 , TRUE)
ps_sub = filter_taxa(ps_sub, function(x) sum(x ) < 30 , TRUE)
ps_sub

result = div_network(ps_sub,num = 6)

edge = result[[1]]
head(edge)
result <- div_culculate(table = result[[3]],distance = 1.1,distance2 = 1.5,distance3 = 1.3,order = FALSE)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
