% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeBuild.R
\name{edgeBuild}
\alias{edgeBuild}
\title{Construct edge files, add weights, positive and negative correlation and other basic information}
\usage{
edgeBuild(cor = cor, plotcord = node)
}
\arguments{
\item{cor}{Correlation matrix}

\item{node}{Node file, containing calculated node coordinates}
}
\value{
edge which contains OTU and its coordinates, the correlation of edge between nodes
}
\description{
Construct edge files, add weights, positive and negative correlation and other basic information
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
netClu = data.frame(ID = row.names(cor),group =rep(1:3,length(row.names(cor)))[1:length(row.names(cor))] )
netClu$group = as.factor(netClu$group)
#Calculate node location
result2 = PolygonRrClusterG(cor = cor,nodeGroup = netClu )
node = result2[[1]]
edge = edgeBuild(cor = cor,plotcord = node)

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
