% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.roles.R
\name{facet.zipi.fast}
\alias{facet.zipi.fast}
\title{Multi-group Zi-Pi plot}
\usage{
facet.zipi.fast(x, zi_cut = 2.5, pi_cut = 0.62, label_key = TRUE)
}
\arguments{
\item{x}{Combined Zi-Pi data containing p, z and group.}

\item{zi_cut}{Zi hub threshold.}

\item{pi_cut}{Pi connector threshold.}

\item{label_key}{Label key-role nodes.}
}
\description{
Multi-group Zi-Pi plot
}
