% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls2.R
\name{filter_OTU_ps}
\alias{filter_OTU_ps}
\title{filter microbiome data}
\usage{
filter_OTU_ps(ps = ps, Top = NULL)
}
\arguments{
\item{ps}{phyloseq abject containg microbiome data}

\item{Top}{number of top abundance tax or OTU}
}
\description{
filter microbiome data
}
\details{
filter_microbiome data
}
\examples{
data(ps)
ps_sub = filter_OTU_ps(ps = ps,Top = 100)
}
