% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls2.R
\name{filter_OTU_ps2}
\alias{filter_OTU_ps2}
\title{filter microbiome data}
\usage{
filter_OTU_ps2(ps = ps, filter = NULL)
}
\arguments{
\item{ps}{phyloseq abject containg microbiome data}

\item{filter}{number or percent of tax or OTU}
}
\description{
filter microbiome data
}
\details{
filter_microbiome data upper
}
\examples{
data(ps)
}
