% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{inputMicro}
\alias{inputMicro}
\title{Quickly import microbiome data and bulding phyloseq object}
\usage{
inputMicro(otu = NULL, tax = NULL, map = NULL, tree = NULL,
  ps = NULL, group = "Group")
}
\arguments{
\item{otu}{dataframe,otu table}

\item{tax}{data.frame, tax table}

\item{map}{matrix or dataframe, including sampleID and groupID;}

\item{tree}{Evolutionary tree}

\item{ps}{a phyloseq object}

\item{group}{column name for groupID.}
}
\value{
a phyloseq object
}
\description{
Quickly import microbiome data and bulding phyloseq object
}
\examples{
data(otu)
data(tax)
data(map)
inputMicro(otu,tax,map,tree,group  = group)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
