% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatCorPlot.R
\name{MatCorPlot}
\alias{MatCorPlot}
\title{Microbial matel text and env correcton with the plot to showing,from science peper.}
\usage{
MatCorPlot(
  env.dat,
  tabOTU,
  distance = "bray",
  method = "metal",
  method.cor = "spearman",
  cor.p = 0.05,
  x = TRUE,
  y = TRUE,
  diag = T,
  sig = TRUE,
  siglabel = FALSE,
  shownum = TRUE,
  numpoint = 22,
  numpoint2 = 21,
  numsymbol = NULL,
  curvature = 0.2,
  lacx = "left",
  lacy = "bottom",
  range = 0.5,
  p.thur = 0.3,
  onlysig = TRUE
)
}
\arguments{
\item{env.dat}{env table}

\item{tabOTU}{OTU table conbined with list}

\item{distance}{The defult, distance = "bray", method of microbiota distance}

\item{method}{method for matel test}

\item{x}{x axis label}

\item{y}{y axis label}

\item{diag}{diag label}

\item{sig}{if show only sig value}

\item{siglabel}{R2 value}

\item{numpoint}{point type 22}

\item{numpoint2}{point type 21}

\item{numsymbol}{point adding  e.g. 27}

\item{lacx}{location of x axis}

\item{lacy}{location of y axis}

\item{range}{curve size}

\item{p.thur}{sig value}

\item{onlysig}{T or F}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{fill}{fill coulor of node}
}
\value{
ggplot object
}
\description{
Microbial matel text and env correcton with the plot to showing,from science peper.
}
\examples{
data(ps)
library(phyloseq)
library(tidyverse)
ps1 = filter_taxa(ps, function(x) sum(x ) > 200 , TRUE);ps1
otu = as.data.frame(t(vegan_otu(ps1)))
mapping = as.data.frame( sample_data(ps1))
tabOTU1 = list(otu1 = otu,otu2 = otu,otu3 = otu)
data(env1)
MatCorPlot(
  env.dat = env1,
 tabOTU = tabOTU1,
  distance = "bray",
  method = "metal",
  x = F,
  y = F,
  diag = T,
  sig = TRUE,
  siglabel = FALSE,
  shownum = TRUE,
  numpoint = NULL,
  numsymbol = 27,
  lacx = "left",
  lacy = "bottom",
  range = 0.5,
  p.thur = 0.3,
  onlysig = F
)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
