% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge16S_ITS.R
\name{merge16S_ITS}
\alias{merge16S_ITS}
\title{merge phyloseq object of 16s and ITS}
\usage{
merge16S_ITS(
  ps16s = ps16s,
  psITS = psITS,
  N16s = 0.001,
  NITS = 0.001,
  scale = TRUE,
  onlygroup = FALSE,
  dat1.lab = "bac",
  dat2.lab = "fun"
)
}
\arguments{
\item{ps16s}{phyloseq Object bacterial, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{psITS}{phyloseq Object fungi, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N16s}{filter OTU tables by abundance.The defult, N=0.02, extract the top 0.02 relative abundance of OTU.}

\item{NITS}{filter OTU tables by abundance.The defult, N=0.02, extract the top 0.02 relative abundance of OTU.}

\item{scale}{transform to reletive abundance}
}
\value{
list which contains OTU correlation matrix
}
\description{
merge phyloseq object of 16s and ITS
}
\examples{
data(ps)
result <- corMicro(ps = ps,N = 0.02,r.threshold=0.6,p.threshold=0.05,method = "pearson")
# extract cor matrix
cor = result[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
