% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Gephi.R
\name{Model_Gephi}
\alias{Model_Gephi}
\title{Model_Gephi:The algorithm imitate Gephi's network layout. while, it saves time and the calculation speed is faster}
\usage{
Model_Gephi(cor = cor, method = "cluster_fast_greedy")
}
\arguments{
\item{cor}{Correlation matrix Clustering Algorithm}

\item{method}{Clustering Algorithm}
}
\value{
list
}
\description{
Enter correlation matrix, calculate network modules, and Calculate the coordinates of the node.
}
\details{
# The algorithm imitate Gephi's network layout. while, it saves time and the calculation speed is faster


By default, returns a list
\itemize{
\item{cluster_fast_greedy: }
\item{cluster_walktrap: }
\item{cluster_edge_betweenness: }
\item{cluster_spinglass: }
}
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
result2 = Model_Gephi(cor = cor)
node = result2[[1]]
node = result2[[2]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
