% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_maptree.R
\name{model_maptree}
\alias{model_maptree}
\title{This is the upgraded version.model_Gephi.2:The algorithm imitate Gephi's network layout. while, it saves time and the calculation speed is faster}
\usage{
model_maptree(cor = cor, method = "cluster_fast_greedy", seed = 12)
}
\arguments{
\item{cor}{Correlation matrix Clustering Algorithm}

\item{method}{Clustering Algorithm}

\item{seed}{Set random seed}
}
\value{
data.frame
}
\description{
Enter correlation matrix, calculate network modules, and Calculate the coordinates of the node.
}
\details{
# The algorithm imitate maptree network layout. while, it saves time and the calculation speed is faster,This is the upgraded version.


By default, returns a list
\itemize{
\item{cluster_fast_greedy: }
\item{cluster_walktrap: }
\item{cluster_edge_betweenness: }
\item{cluster_spinglass: }
}
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
result2 <- model_maptree(cor = cor,
method = "cluster_fast_greedy",
seed = 12
)
node = result2
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
