% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_maptree2.R
\name{model_maptree2}
\alias{model_maptree2}
\title{This is the upgraded version.model_Gephi.2:The algorithm imitate Gephi's network layout. while, it saves time and the calculation speed is faster}
\usage{
model_maptree2(cor = cor, method = "cluster_fast_greedy", seed = 12)
}
\arguments{
\item{cor}{Correlation matrix Clustering Algorithm}

\item{method}{Clustering Algorithm}

\item{seed}{Set random seed}
}
\value{
data.frame
}
\description{
Enter correlation matrix, calculate network modules, and Calculate the coordinates of the node.
}
\details{
# The algorithm imitate maptree network layout. while, it saves time and the calculation speed is faster,This is the upgraded version.


By default, returns a list
\itemize{
\item{cluster_fast_greedy: }
\item{cluster_walktrap: }
\item{cluster_edge_betweenness: }
\item{cluster_spinglass: }
}
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 100,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
result2 <- model_maptree(cor = cor,
method = "cluster_fast_greedy",
seed = 12
)
node = result2
}
\references{
Tao Wen#, Penghao Xie#, Shengdie Yang, Guoqing Niu, Xiaoyu Liu, Zhexu Ding, Chao Xue, Yong-Xin Liu *, Qirong Shen, Jun Yuan*
ggClusterNet: an R package for microbiome network analysis and modularity-based multiple network layouts
iMeta 2022,DOI: \url{doi: 10.1002/imt2.32}
}
\author{
Contact: Tao Wen \email{taowen@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn} yongxin liu \email{yxliu@genetics.ac.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
