% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.roles.R
\name{module.roles}
\alias{module.roles}
\title{Fast module roles}
\usage{
module.roles(
  comm_graph,
  membership = NULL,
  role_weights = c("degree", "strength"),
  zi_cut = 2.5,
  pi_cut = 0.62
)
}
\description{
Uses an existing vertex attribute "module" when membership is not supplied.
}
