% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2.R
\name{network.2}
\alias{network.2}
\title{Microbial related network}
\usage{
network.2(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  N = 0,
  big = FALSE,
  select_layout = FALSE,
  layout_net = "model_maptree",
  r.threshold = 0.6,
  p.threshold = 0.05,
  method = "spearman",
  label = FALSE,
  lab = "elements",
  group = "Group",
  path = "./",
  fill = "Phylum",
  size = "igraph.degree",
  scale = TRUE,
  zipi = FALSE,
  clu_method = "cluster_fast_greedy",
  step = 100,
  yourmem = theme_void(),
  ncol = 3,
  nrow = 1,
  R = 10,
  ncpus = 1
)
}
\arguments{
\item{ps}{phyloseq Object, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N}{filter OTU tables by abundance.The defult, N=0, extract the top N number relative abundance of OTU. e.g 100}

\item{big}{TRUE or FALSE the number of micro data was so many (> 300),you can chose TREU}

\item{select_layout}{TURE or FALSE}

\item{layout_net}{select layout from ggClusterNet}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{path}{save path of all of network analyse.}

\item{fill}{fill coulor of node}

\item{size}{node size}

\item{zipi}{zipi Calculation}

\item{step}{Random network sampling times}

\item{R}{repeat number of p value calculate}

\item{ncpus}{number of cpus used for sparcc}

\item{lay}{layout which network show}
}
\value{
list which contains OTU correlation matrix
}
\description{
Microbial related network
}
\examples{
data(ps)
result = network (ps = ps,N = 0.001,r.threshold=0.6,p.threshold=0.05,label = FALSE,path = path ,zipi = TRUE)
result[[1]]
result[[2]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
