% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.pip.R
\name{network.pip}
\alias{network.pip}
\title{Microbial related network}
\usage{
network.pip(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  N = 0,
  ra = NULL,
  order = NULL,
  big = TRUE,
  select_layout = FALSE,
  layout_net = "model_maptree2",
  r.threshold = 0.6,
  p.threshold = 0.05,
  maxnode = 2,
  method = "spearman",
  label = FALSE,
  lab = "elements",
  group = "Group",
  fill = "Phylum",
  size = "igraph.degree",
  zipi = FALSE,
  clu_method = "cluster_fast_greedy",
  step = 100,
  ram.net = FALSE,
  R,
  ncpus
)
}
\arguments{
\item{ps}{phyloseq Object, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N}{filter OTU tables by abundance.The defult, N=0, extract the top N number relative abundance of OTU. e.g 100}

\item{big}{TRUE or FALSE the number of micro data was so many (> 300),you can chose TREU}

\item{select_layout}{TURE or FALSE}

\item{layout_net}{select layout from ggClusterNet}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{fill}{fill coulor of node}

\item{size}{node size}

\item{zipi}{zipi Calculation}

\item{step}{Random network sampling times}

\item{R}{repeat number of p value calculate}

\item{ncpus}{number of cpus used for sparcc}

\item{lay}{layout which network show}

\item{path}{save path of all of network analyse.}
}
\value{
list which contains OTU correlation matrix
}
\description{
Microbial related network
}
\examples{
data(ps)
result = network.pip(ps = ps,N = 100,r.threshold=0.6,big = T,
                   select_layout = T,
                   p.threshold=0.05,label = FALSE,zipi = F)
result[[1]]
result[[2]]
}
\references{
Tao Wen#, Penghao Xie#, Shengdie Yang, Guoqing Niu, Xiaoyu Liu, Zhexu Ding, Chao Xue, Yong-Xin Liu *, Qirong Shen, Jun Yuan*
ggClusterNet: an R package for microbiome network analysis and modularity-based multiple network layouts
iMeta 2022,DOI: \url{doi: 10.1002/imt2.32}
}
\author{
Contact: Tao Wen \email{taowen@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn} yongxin liu \email{yxliu@genetics.ac.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
