% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkplot.R
\name{network}
\alias{network}
\title{Microbial related network}
\usage{
network(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  N = 0.001,
  r.threshold = 0.6,
  p.threshold = 0.05,
  method = "pearson",
  label = FALSE,
  group = "Group",
  lay = "fruchtermanreingold",
  path = "./",
  fill = "Phylum",
  size = "igraph.degree",
  scale = TRUE,
  zipi = FALSE,
  clu_method = "cluster_fast_greedy",
  step = 100
)
}
\arguments{
\item{ps}{phyloseq Object, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N}{filter OTU tables by abundance using otu table with all samples.The defult, N=0.001, extract the top 0.001 relative abundance of OTU.}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{lay}{layout which network show}

\item{path}{save path of all of network analyse.}

\item{fill}{fill coulor of node}

\item{size}{node size}

\item{zipi}{zipi Calculation}

\item{step}{Random network sampling times}
}
\value{
list which contains OTU correlation matrix
}
\description{
Microbial related network
}
\examples{
data(ps)
result = network (ps = ps,N = 0.001,r.threshold=0.6,p.threshold=0.05,label = FALSE,path = path ,zipi = TRUE)
result[[1]]
result[[2]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
