% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_properties.R
\name{node_properties}
\alias{node_properties}
\title{Compute network node attributes}
\usage{
node_properties(igraph, outdir)
}
\arguments{
\item{igraph}{newwork}
}
\value{
list
}
\description{
Compute network node attributes
}
\examples{
data(igraph)
node_properties(igraph)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
