% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputgephi.R
\name{outputgephi}
\alias{outputgephi}
\title{use the phylsoeq object and cor matrix building Gephi input network file contain edge and node}
\usage{
outputgephi(ps = ps, cor = cor)
}
\arguments{
\item{ps}{phyloseq abject}

\item{cor}{Correlation matrix}
}
\value{
result2 Which contains 2 lists.Result2[[1]], consists of OTU and its corresponding coordinates.
Result2[[2]], consists of the network center coordinates of each group
}
\description{
use the phylsoeq object and cor matrix building Gephi input network file contain edge and node
}
\examples{
data
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
gephi <- outputgephi(ps = result[[3]],cor =  cor)
edge_Gephi <- gephi[[1]]
head(edge_Gephi)
node_Gephi <- gephi[[2]]
head(node_Gephi)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
