% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonMaptreeG.R
\name{PolygonMaptreeG}
\alias{PolygonMaptreeG}
\title{Correlation matrix calculates groups according to the degree of network module speech.}
\usage{
PolygonMaptreeG(cor = cor, nodeGroup = gru, seed = 12)
}
\arguments{
\item{cor}{Correlation matrix}

\item{nodeGroup}{Microbiome grouping table}

\item{seed}{set random seed,Default 12}
}
\value{
data table.
}
\description{
Enter correlation matrix, calculate network modules, and generate groups.
}
\details{
# This is the function named 'modulGroup'# which Enter the correlation matrix. There are four module clustering algorithms inside, which can calculate the modularity and then group according to the modularity.


By default, returns table, contain node and group imformation
The available method to culculate Degree of modularity include the following:
\itemize{
\item{cluster_fast_greedy: }
\item{cluster_walktrap: }
\item{cluster_edge_betweenness: }
\item{cluster_spinglass: }
}
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 100,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
PolygonMaptreeG(cor = cor,nodeGroup = gru,seed = 12)

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
