% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_merge_Tax.R
\name{ps_merge_Tax}
\alias{ps_merge_Tax}
\title{Stacked histograms add error bars, marked by significance}
\usage{
ps_merge_Tax(ps = ps, j = "Phylum", Top = 10)
}
\arguments{
\item{ps}{a object of phyloseq}

\item{j}{rank of tax}

\item{Top}{top number of number selected}

\item{errbar}{add a distinctive label, TRUE or FEASE could be selected}

\item{result}{output from aovMcomper or KwWlx. You can also import result calculated from other software (a data frame)}

\item{add_abc}{add sig abc T or F}
}
\value{
list
}
\description{
Stacked histograms add error bars, marked by significance
}
\examples{
ps_merge_Tax (ps = ps,j = "Phylum",Top = 10)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} PengHao Xie \email{2019103106@njau.edu.cn}
}
