% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stable.R
\name{rand.remov.once}
\alias{rand.remov.once}
\title{id.rm' species; not consider the sequential effects of extinction of}
\usage{
rand.remov.once(netRaw, rm.percent, sp.ra, abundance.weighted = T)
}
\description{
id.rm' species; not consider the sequential effects of extinction of
}
