% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomClusterG.R
\name{randomClusterG}
\alias{randomClusterG}
\title{Construct a network layout. Randomly distributed module positions}
\usage{
randomClusterG(cor = cor, nodeGroup = netClu)
}
\arguments{
\item{cor}{Correlation matrix}

\item{nodeGroup}{Classification information of network nodes}
}
\value{
list
}
\description{
Construct a network layout. Randomly distributed module positions
}
\examples{
data
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# building the node group
netClu = data.frame(ID = row.names(cor),group =rep(1:3,length(row.names(cor)))[1:length(row.names(cor))] )
netClu$group = as.factor(netClu$group)
result2 = randomClusterG (cor = cor,nodeGroup =netClu ）
node = result2[[1]]


}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
