% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls2.R
\name{scale_micro}
\alias{scale_micro}
\title{scaling microbiome data}
\usage{
scale_micro(ps, method = "rela")
}
\arguments{
\item{ps}{phyloseq abject containg microbiome data}

\item{method}{could be selected byy rela, sampling, log,TMM, et al}
}
\description{
will use many method for microbiome data scaling
}
\details{
scale_microbiome data
}
\examples{
scale_micro(ps = ps,method = "rela")
}
