% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_glom_wt.R
\name{tax_glom_wt}
\alias{tax_glom_wt}
\title{Combine microbiome data by classification level}
\usage{
tax_glom_wt(ps = ps, ranks = "Phylum")
}
\arguments{
\item{ps}{phyloseq object}

\item{ranks}{tax of microbiome data,could be one of Phylum, Order, Class, Genus ta al.}
}
\value{
S4 phyloseq abject
}
\description{
Combine microbiome data by classification level
}
\examples{
tax_glom_wt(ps = ps,ranks = "Genus")
tax_glom_wt(ps = ps,ranks = "Class")
tax_glom_wt(ps = ps,ranks = "Order")
tax_glom_wt(ps = ps,ranks = "Phylum")
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
