% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{vegan_tax}
\alias{vegan_tax}
\title{transform a tax object from phyloseq object to matrix}
\usage{
vegan_tax(physeq)
}
\arguments{
\item{physeq}{a phyloseq object from R package phyloseq}
}
\value{
matrix
}
\description{
transform a tax object from phyloseq object to matrix
}
\examples{
tax = vegan_tax(ps)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
