% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtifClusterG.R
\name{ArtifCluster}
\alias{ArtifCluster}
\title{Manually specify coordinates to display the network}
\usage{
ArtifCluster(cor = cor, nodeGroup = netClu, r = r, da = da)
}
\arguments{
\item{cor}{Correlation matrix}

\item{nodeGroup}{Classification information of network nodes.Group according to actual requirements, see example}

\item{r}{Radius of each submodule}

\item{da}{The coordinates of each submodule}
}
\value{
list which contains node position coordinates
}
\description{
Manually specify coordinates to display the network
}
\examples{
data(ps)
result = corMicro (ps = ps,N = 100,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# Extract taxonomy table
ps_net = result[[3]]
# set group of experimental
netClu = data.frame(ID = row.names(tax_table),group =rep(1:5,length(row.names(tax_table)))[1:length(row.names(tax_table))] )
netClu$group = as.factor(netClu$group)
tax_table = ps_net \%>\% vegan_tax() \%>\%
as.data.frame()
xs = as.data.frame(table(netClu$group))
# set the radicus
r = rep(4,length(xs$Freq))
# Set the coordinates manually
ax1 = c(120,0)
ax2 = c(130,-30)
ax3 = c(140,-70)
ax4 = c(130,-110)
ax5 = c(120,-140)
da = rbind(ax1,ax2,ax3,ax4,ax5)
# Calculate network layout
result2 = ArtifCluster(cor = cor,nodeGroup =netClu,r = r,da =da)
node = result2[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
