% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetworkplot.R
\name{corBionetwork}
\alias{corBionetwork}
\title{Microbial related network}
\usage{
corBionetwork(otu = NULL, tax = NULL, map = NULL, ps = NULL,
  N = 0.001, r.threshold = 0.6, p.threshold = 0.05, label = FALSE,
  group = "Group", env = NULL, envGroup = NULL,
  method = "spearman", layout = "fruchtermanreingold", path = "./",
  fill = "Phylum", size = "igraph.degree", scale = TRUE,
  bio = FALSE, zipi = FALSE, step = 100, width = 8, height = 6,
  allnode = TRUE)
}
\arguments{
\item{ps}{phyloseq Object, contains OTU tables, tax table and map table, represented sequences,phylogenetic tree.}

\item{N}{filter OTU tables by abundance.The defult, N=0.001, extract the top 0.001 relative abundance of OTU.}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{env}{Environmental factor index table which do network analysis with the microbiome.}

\item{envGroup}{group of env table.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{path}{save path of all of network analyse.}

\item{fill}{fill coulor of node.}

\item{size}{node size.}

\item{scale}{Whether relative abundance standardization is required.}

\item{bio}{Do you need to do a binary network.}

\item{zipi}{zipi Calculation.}

\item{step}{Random network sampling times.}

\item{width}{Save the width of the picture settings.}

\item{height}{Save the height of the picture setting.}

\item{lay}{layout which network show.}
}
\value{
list which contains OTU correlation matrix
}
\description{
Microbial related network
}
\examples{
data(ps)
result = network (ps = ps,N = 0.001,r.threshold=0.6,p.threshold=0.05,label = FALSE,path = path ,zipi = TRUE)
result[[1]]
result[[2]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
