% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_properties.R
\name{net_properties}
\alias{net_properties}
\title{Use the layout in the sne package to calculate the visual layout of the network}
\usage{
net_properties(igraph)
}
\arguments{
\item{igraph}{network}
}
\value{
list
}
\description{
Use the layout in the sne package to calculate the visual layout of the network
}
\examples{
# Import network data
data(igraph)
# Computing network related properties锛歵he total number of edges, the number of positive correlation edges, the number of negative correlation edges, the number of vertices...
reslt = net_properties(igraph)
# export file
write.csv(reslt,"netproperties.csv")

}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
