% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeadd.R
\name{nodeadd}
\alias{nodeadd}
\title{Use otu forms and annotation files to annotate microbial network nodes}
\usage{
nodeadd(plotcord = node, otu_table = otu_table,
  tax_table = tax_table)
}
\arguments{
\item{ps}{phyloseq Object, could contains OTU tables, tax table and map table, represented sequences.}

\item{N}{number object, filter OTU tables by abundance.}

\item{r.threshold}{Correlation threshold}

\item{p.threshold}{Significance threshold}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}
}
\value{
list
}
\description{
Use otu forms and annotation files to annotate microbial network nodes
}
\examples{
nodes = nodeadd(plotcord =node,otu_table = otu_table,tax_table = tax_table)
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
