% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomClusterG.R
\name{randomClusterG}
\alias{randomClusterG}
\title{Construct a network layout. Randomly distributed module positions}
\usage{
randomClusterG(cor = cor, nodeGroup = netClu)
}
\arguments{
\item{cor}{Correlation matrix}

\item{nodeGroup}{Classification information of network nodes}
}
\value{
result2 Which contains 2 data.frame. Result2[[1]], consists of OTU and its corresponding coordinates.
result2[[2]], consists of the network center coordinates of each group
}
\description{
Construct a network layout. Randomly distributed module positions
}
\examples{
data
data(ps)
result = corMicro (ps = ps,N = 0.02,r.threshold=0.8,p.threshold=0.05,method = "pearson")
#Extract correlation matrix
cor = result[[1]]
# Extract tax table for grouping
ps_net = result[[3]]
vegan_tax <-  function(physeq){
tax <-  tax_table(physeq)

return(as(tax,"matrix"))
}
tax_table = as.data.frame(vegan_tax(ps_net))
group = as.data.frame(tax_table)
group$ID = row.names(group)
netClu = data.frame(ID = row.names(group),group = group$Phylum)
# Calculate the layout of netwok, extract the coordinate of node
result2 = randomClusterG (cor = cor,nodeGroup =netClu )
node = result2[[1]]


}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn} Penghao Xie \email{2019103106@njau.edu.cn}
}
