% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZiPiPlot.R
\name{ZiPiPlot}
\alias{ZiPiPlot}
\title{Use the layout in the sne package to calculate the visual layout of the network}
\usage{
ZiPiPlot(igraph = igraph, method = "cluster_fast_greedy")
}
\arguments{
\item{method}{method to culculate Degree of modularity}

\item{cor}{Correlation matrix}
}
\value{
list
}
\description{
Use the layout in the sne package to calculate the visual layout of the network
}
\examples{
data(igraph)
res = ZiPiPlot(igraph = igraph,method = "cluster_fast_greedy")
res[[1]]
}
\references{
Yuan J, Zhao J, Wen T, Zhao M, Li R, Goossens P, Huang Q, Bai Y, Vivanco JM, Kowalchuk GA, Berendsen RL, Shen Q
Root exudates drive the soil-borne legacy of aboveground pathogen infection
Microbiome 2018,DOI: \url{doi: 10.1186/s40168-018-0537-x}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn} Jun Yuan \email{junyuan@njau.edu.cn}
}
